# src/COP.API/Controllers/WearableApi.cs
/*
 * MONICA COP API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using COP.API.Attributes;
using COP.API.Models;
using COP.API.DatabaseInterface;

namespace COP.API.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class WearableApiController : Controller
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>add a new wearable</remarks>
        /// <param name="name">name of wearable</param>
        /// <response code="200">Success</response>
        /// <response code="0">Error</response>
        [HttpPost]
        [Route("/wearable")]
        [ValidateModelState]
        [SwaggerOperation("WearablePost")]
        [SwaggerResponse(statusCode: 200, type: typeof(GeneralResponse), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "Error")]
        public virtual IActionResult WearablePost([FromBody]Wearable name, [FromHeader][Required()]string Authorization)
        {

            if (!ModelState.IsValid)
            {
                var error = ModelState.SelectMany(x => x.Value.Errors).First();
                if (error.ErrorMessage != null && error.ErrorMessage != String.Empty)
                {
                    return BadRequest(error.ErrorMessage);
                }
                else if (error.Exception?.Message != null)
                {
                    return BadRequest("Faulty input");
                }
                else
                    return BadRequest(ModelState);

            }
            if (Authorization != settings.testToken)
                return BadRequest("Not allowed");
            long newWearableID = 0;
            string errorMessage = "";
            try
            {
                DatabaseInterface.DBWearable dBWearable = new DatabaseInterface.DBWearable();
                if (!dBWearable.AddWearable(name.PersonId, name.ThingId, ref errorMessage, ref newWearableID))
                    return BadRequest("Internal Server Error:" + errorMessage);

            }
            catch (Exception e)
            {
                return BadRequest("Internal Server Error:" + e.Message);
            }


            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GeneralResponse));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ErrorResponse));


            string exampleJson = null;
            exampleJson = "{\n  \"success\" : true,\n  \"newid\" : " + newWearableID.ToString() + "\n}";

            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GeneralPostResponse>(exampleJson)
            : default(GeneralPostResponse);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>delete wearable by id</remarks>
        /// <param name="id"></param>
        /// <response code="200">Success</response>
        /// <response code="0">Error</response>
        [HttpDelete]
        [Route("/wearables/{id}")]
        [ValidateModelState]
        [SwaggerOperation("WearablesIdDelete")]
        [SwaggerResponse(statusCode: 200, type: typeof(GeneralResponse), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "Error")]
        public virtual IActionResult WearablesIdDelete([FromRoute][Required]string id, [FromHeader][Required()]string Authorization)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GeneralResponse));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ErrorResponse));
            if (Authorization != settings.testToken)
                return BadRequest("Not allowed");
            string exampleJson = null;
            exampleJson = "{\n  \"success\" : true,\n  \"description\" : \"description\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GeneralResponse>(exampleJson)
            : default(GeneralResponse);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>get wearable by id</remarks>
        /// <param name="id"></param>
        /// <response code="200">Success</response>
        /// <response code="0">Error</response>
        [HttpGet]
        [Route("/wearables/{id}")]
        [ValidateModelState]
        [SwaggerOperation("WearablesIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(Wearable), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "Error")]
        public virtual IActionResult WearablesIdGet([FromRoute][Required]string id, [FromHeader][Required()]string Authorization)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Wearable));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ErrorResponse));
            if (Authorization != settings.testToken)
                return BadRequest("Not allowed");
            string exampleJson = null;
            exampleJson = "{\n  \"personId\" : 0,\n  \"thingId\" : 6,\n  \"timestamp\" : \"2000-01-23T04:56:07.000+00:00\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Wearable>(exampleJson)
            : default(Wearable);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>update Wearable</remarks>
        /// <param name="id"></param>
        /// <param name="updates"></param>
        /// <response code="200">Success</response>
        /// <response code="0">Error</response>
        [HttpPut]
        [Route("/wearables/{id}")]
        [ValidateModelState]
        [SwaggerOperation("WearablesIdPut")]
        [SwaggerResponse(statusCode: 200, type: typeof(GeneralResponse), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "Error")]
        public virtual IActionResult WearablesIdPut([FromRoute][Required]string id, [FromBody]Wearable updates, [FromHeader][Required()]string Authorization)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GeneralResponse));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ErrorResponse));
            if (Authorization != settings.testToken)
                return BadRequest("Not allowed");
            string exampleJson = null;
            exampleJson = "{\n  \"success\" : true,\n  \"description\" : \"description\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GeneralResponse>(exampleJson)
            : default(GeneralResponse);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>get wearable by id</remarks>
        /// <param name="id"></param>
        /// <response code="200">Success</response>
        /// <response code="0">Error</response>
        [HttpGet]
        [Route("/wearables/getPerson/{wbid}")]
        [ValidateModelState]
        [SwaggerOperation("WearablesPersonIdGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(int), description: "Success")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "Error")]
        public virtual IActionResult WearablesPersonIdGet([FromRoute][Required]string wbid, [FromHeader][Required()]string Authorization)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Wearable));

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default(ErrorResponse));
            if (!ModelState.IsValid)
            {
                var error = ModelState.SelectMany(x => x.Value.Errors).First();
                if (error.ErrorMessage != null && error.ErrorMessage != String.Empty)
                {
                    return BadRequest(error.ErrorMessage);
                }
                else if (error.Exception?.Message != null)
                {
                    return BadRequest("Faulty input");
                }
                else
                    return BadRequest(ModelState);

            }
            if (Authorization != settings.testToken)
                return BadRequest("Not allowed");
            List<COP.API.Models.Wearable> results = null;
            string errorMessage = "";
            try
            {
                DatabaseInterface.DBWearable dBWearable = new DatabaseInterface.DBWearable();
                if (!dBWearable.WearablesPersonIdGet(wbid, ref errorMessage, ref results))
                    return BadRequest("Internal Server Error:" + errorMessage);

            }
            catch (Exception e)
            {
                return BadRequest("Internal Server Error:" + e.Message);
            }
       
         
            //TODO: Change the data returned
            return new ObjectResult(results[0].PersonId);
        }
    }
}