# src/COP.API/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.1.0",
    "title" : "MONICA COP API"
  },
  "host" : "portal.monica-cloud.eu",
  "basePath" : "/cop/api",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/wearable" : {
      "post" : {
        "tags" : [ "Wearable" ],
        "description" : "add a new wearable",
        "parameters" : [ {
          "in" : "body",
          "name" : "name",
          "description" : "name of wearable",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Wearable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/wearables/{id}" : {
      "get" : {
        "tags" : [ "Wearable" ],
        "description" : "get wearable by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Wearable"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Wearable" ],
        "description" : "update Wearable",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "updates",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Wearable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Wearable" ],
        "description" : "delete wearable by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/GeneralResponse"
            }
          },
          "default" : {
            "description" : "Error",
            "schema" : {
              "$ref" : "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Wearable" : {
      "type" : "object",
      "properties" : {
        "personId" : {
          "type" : "integer",
          "description" : "Person id"
        },
        "thingId" : {
          "type" : "integer",
          "description" : "Thing ID"
        },
        "timestamp" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "time of wearable connection (e.g. '2016-06-17T15:28:34Z' (RFC 3339, ISO 8601))"
        }
      },
      "example" : {
        "personId" : 0,
        "thingId" : 6,
        "timestamp" : "2000-01-23T04:56:07.000+00:00"
      }
    },
    "GeneralResponse" : {
      "type" : "object",
      "required" : [ "description", "success" ],
      "properties" : {
        "success" : {
          "type" : "boolean",
          "description" : "returns 1 if successful"
        },
        "description" : {
          "type" : "string",
          "description" : "a short comment"
        }
      },
      "example" : {
        "success" : true,
        "description" : "description"
      }
    },
    "ErrorResponse" : {
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}